@@ -98,7 +98,20 @@ func TestGCS(t *testing.T) {
 		t.Fatalf("failed to create gcs storage: %s", err)
 	}
 
-	storage.RunStorageTests(t, stg, storage.GCS)
+	storage.RunStorageBaseTests(t, stg, storage.GCS)
+	storage.RunStorageAPITests(t, stg)
+	storage.RunSplitMergeMWTests(t, stg)
+
+	t.Run("with downloader", func(t *testing.T) {
+		stg, err := NewWithDownloader(opts, "node", nil, 0, 0, 0)
+		if err != nil {
+			t.Fatalf("failed to create gcs storage: %s", err)
+		}
+
+		storage.RunStorageBaseTests(t, stg, storage.GCS)
+		storage.RunStorageAPITests(t, stg)
+		storage.RunSplitMergeMWTests(t, stg)
+	})
 
 	t.Run("Delete fails", func(t *testing.T) {
 		name := "not_found.txt"
