@@ -73,7 +73,20 @@ func TestS3(t *testing.T) {
 		t.Fatalf("failed to create s3 storage: %s", err)
 	}
 
-	storage.RunStorageTests(t, stg, storage.S3)
+	storage.RunStorageBaseTests(t, stg, storage.S3)
+	storage.RunStorageAPITests(t, stg)
+	storage.RunSplitMergeMWTests(t, stg)
+
+	t.Run("with downloader", func(t *testing.T) {
+		stg, err := NewWithDownloader(opts, "node", nil, 0, 0, 0)
+		if err != nil {
+			t.Fatalf("failed to create s3 storage: %s", err)
+		}
+
+		storage.RunStorageBaseTests(t, stg, storage.S3)
+		storage.RunStorageAPITests(t, stg)
+		storage.RunSplitMergeMWTests(t, stg)
+	})
 
 	t.Run("default SDKLogLevel for invalid value", func(t *testing.T) {
 		logLvl := SDKLogLevel("invalid", nil)
